+++ block-os/src/editor/extensions/index.ts
@@ -3,3 +3,14 @@ export { BlockReference } from './blockReference'
 export { SourceBlock } from './sourceBlock'
 export { searchBlocks, createSuggestionPlugin } from './suggestion'
 export type { SuggestionItem } from './suggestion'
+export {
+  createInlineAIPlugin,
+  inlineAIPluginKey,
+  startInlineAIReplace,
+  updateInlineAIContent,
+  confirmInlineAIReplace,
+  discardInlineAIReplace,
+  cancelActiveInlineAI,
+  hasActiveInlineAI,
+} from './inlineAI'
+export type { InlineAIState } from './inlineAI'